requirements.txt
flask==2.3.2
scapy==2.5.1
python-dotenv==1.0.0